+++ CHANGELOG.md
@@ -14,6 +14,7 @@ date.
 * Added `accounthist` tools allows you to scan and read accounts `dfuseeos tools accounthist read ...` `dfuseeos tools accounthist scan ...`
 * Flag `--search-router-truncation-low-block-num` to make the router aware of lower-block-truncation and serve requests accordingly
 * Flag `--mindreader-oneblock-suffix` that mindreaders can each write their own file per block without competing for writes. https://github.com/dfuse-io/dfuse-eosio/issues/140
+* Flag `--eosws-disabled-messages` a comma separated list of ws messages to disable. 
 
 ### Changed